+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Implement virtual controls
 - Cab Debugger can connect
 - Only show last 1000 logs by default
+- Fix raw events
+- Fix unique joystick IDs
 
 ## 1.3.6
 - Fix bug when saving calibration files